requirements.txt
appdirs==1.4.3
apturl==0.5.2
asgiref==3.7.2
attrs==24.2.0
autobahn==23.1.2
Automat==24.8.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.17.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
constantly==23.10.4
contourpy==1.0.7
cryptography==43.0.1
cupshelpers==1.0
cycler==0.11.0
daphne==4.1.2
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.8
django-cors-headers==4.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
folder-color-common==0.0.88
fonttools==4.39.3
future==0.18.2
httplib2==0.14.0
hyperlink==21.0.0
idna==2.8
importlib-metadata==1.5.0
importlib-resources==5.12.0
incremental==24.7.2
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.7.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.24.3
oauthlib==3.1.0
olefile==0.46
packaging==23.1
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psutil==5.5.1
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycairo==1.16.2
pycparser==2.22
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==24.2.1
pyparsing==3.0.9
pyRFC3339==1.1
python-apt==2.0.1+zorin1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==24.1.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.4
tomli==2.0.1
twisted==24.7.0
txaio==23.1.1
typing-extensions==4.9.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.0.17
wadllib==1.3.3
xkit==0.0.0
zipp==3.15.0
zope.interface==7.0.3
zorin-appearance==4.1
zorin-connect==1.2
zorin-exec-guard==1.5